requirements.txt
# Original dependencies
langchain==0.2.16
langchain-core==0.2.39
langchain-google-genai==1.0.10
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3
python-dotenv==1.0.1
pinecone-client==5.0.1

# LangGraph addition
langgraph>=0.2.0

# Optional: For fallback LLM
langchain-groq>=0.1.0